05.SingleAgent/requirements.txt
fastapi
uvicorn
openai
pydantic
streamlit
requests
python-dotenv
